+++ src/main/java/colorful/starbucks/batch/memberlevel/MemberLevelScheduler.java
@@ -1,4 +1,4 @@
-package colorful.starbucks.batch;
+package colorful.starbucks.batch.memberlevel;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
@@ -22,7 +22,8 @@ public class MemberLevelScheduler {
     private final JobLauncher jobLauncher;
     private final JobRegistry jobRegistry;
 
-    @Scheduled(cron = "0/10 * * * * *")
+    @Scheduled(cron = "0 0 0 * * *")
+
     public void runJob() {
         String time = LocalDateTime.now().toString();
